backend/requirements.txt
fastapi
uvicorn
supabase
bcrypt
python-jose
python-dotenv
email-validator